+++ src/Mapping/Contract/MappingServiceInterface.php
@@ -5,7 +5,6 @@ namespace Heptacom\HeptaConnect\Core\Mapping\Contract;
 
 use Heptacom\HeptaConnect\Core\Mapping\Exception\MappingNodeAreUnmergableException;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\Contract\MappingInterface;
-use Heptacom\HeptaConnect\Portal\Base\Mapping\MappingComponentCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\MappingNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 
@@ -38,8 +37,6 @@ interface MappingServiceInterface
         PortalNodeKeyInterface $portalNodeKey,
         array $externalIds
     ): iterable;
-
-    public function ensurePersistence(MappingComponentCollection $mappingComponentCollection): void;
 
     public function save(MappingInterface $mapping): void;